clean up and organize extension and code action provider like a sane person.
